Refactor pytest into sender and receiver scripts
